fix: improve license file generation
