+++ app/package.json
@@ -2,6 +2,7 @@
   "name": "gover",
   "version": "0.0.0",
   "private": true,
+  "license": "SEE LICENSE IN LICENSE.md",
   "scripts": {
     "start:staff": "VITE_APP_MODE=staff vite --config vite.config.staff.js",
     "start:customer": "VITE_APP_MODE=customer vite --config vite.config.customer.js",
